+++ src/main.rs
@@ -16,21 +16,19 @@ extern crate rusoto_sqs;
 extern crate serde_json;
 extern crate serde;
 
-use std::io::Read;
+mod http;
+
 use std::str;
 use std::collections::{HashMap};
 use std::env;
 
 use config::{ConfigError, Config, Value, File};
 use failure::Error;
-use rocket::{Request, Data, Rocket};
-use rocket::data::{self, FromData};
-use rocket::http::{Status};
-use rocket::Outcome::*;
-use rocket_contrib::Json;
 use rusoto_core::{Region};
 //use rusoto_sns::{Sns, SnsClient, GetTopicAttributesInput};
 //use rusoto_sqs::{Sqs, SqsClient, ListQueuesRequest};
+
+use http::create_rocket;
 
 //TODO: divide into sane components.
 
@@ -53,11 +51,6 @@ enum MegaphoneError {
 // ==== PubSub
 // TODO: create new topic if not present?
 
-
-// Version information from command line.
-struct VersionInput {
-    value: String,
-}
 // ==== mysql
 
 #[derive(Debug, Deserialize)]
@@ -332,100 +325,6 @@ impl AwsService {
     }
 }
 
-// ==== REST
-
-impl FromData for VersionInput {
-    type Error = String;
-
-    fn from_data(_req: &Request, data: Data) -> data::Outcome<Self, String> {
-        let mut string = String::new();
-        if let Err(e) = data.open().read_to_string(&mut string) {
-            return Failure((Status::InternalServerError, format!("{:?}", e)));
-        }
-
-        // TODO Validate the version info
-
-        Success(VersionInput {
-            value: string
-        })
-    }
-}
-
-
-// Generic Response
-#[derive(Serialize)]
-struct MResponse {
-    status: u8,
-    error_code: u8,
-    error: String,
-    body: String,
-}
-
-impl Default for MResponse {
-    fn default() -> MResponse {
-        MResponse {
-            status: 200,
-            error_code: 0,
-            error: String::from(""),
-            body: String::from(""),
-        }
-    }
-}
-
-/*
-impl Serialize for MResponse {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-        where S: Serializer
-    {
-        let mut s = serializer.serialize_struct("MResponse", 3)?;
-        s.serialize_field("status", &self.status)?;
-        if &self.error_code > &u8::from(0) {
-            s.serialize_field("error_code", &self.error_code)?;
-            s.serialize_field("error", &self.error)?;
-        }
-        s.end()
-    }
-}
-*/
-
-// REST Functions
-#[post("/v1/rtu/<broadcaster_id>/<collection_id>", data = "<version>")]
-fn accept(broadcaster_id: String, collection_id: String, version: VersionInput) -> Json<MResponse>{
-    /// Set a version for a broadcaster / collection
-
-    println!("broadcaster: {:?}\n\tcollection: {:?}\n\tversion:{:?}",
-        broadcaster_id, collection_id, version.value
-    );
-
-    // TODO: Validate auth cookie
-    // TODO: Validate broadcaster & collection; create SenderID
-    // TODO: publish version change / update local table.
-    Json(MResponse {..Default::default()})
-}
-
-/* Dump the current table */
-#[get("/v1/rtu")]
-fn dump() -> Json<MResponse> {
-    /// Dump the nodes current version info table.
-    // TODO: dump the local table of senderID -> version data.
-    let settings = Settings::new().expect("Could not get settings");
-    let database = Database::new(settings.database);
-    let snapshot = format!("{:?}", database.snapshot());
-    let response = MResponse {
-        body: snapshot,
-        ..Default::default()
-    };
-    Json(response)
-}
-
-// TODO: Database handler.
-// TODO: PubSub handler.
-// TODO: HTTP Error Handlers  https://rocket.rs/guide/requests/#error-catchers
-
-fn create_rocket() -> Rocket {
-    rocket::ignite().mount("/", routes![accept, dump])
-}
-
 /*
     We need to use SNS for fan out broadcast of updates to the other nodes. Each node will also
     have an SQS queue that it will have to subscribe to the SNS fan-out. The node reads pending
@@ -447,22 +346,4 @@ fn main() {
     let database = Database::new(config.database);
 
     create_rocket().launch();
-}
-
-#[cfg(test)]
-mod test {
-    use super::create_rocket;
-    use rocket::local::Client;
-    use rocket::http::Status;
-    use serde_json;
-
-    #[test]
-    fn hello_world() {
-        let client = Client::new(create_rocket()).expect("valid rocket instance");
-        let mut response = client.get("/v1/rtu").dispatch();
-        assert_eq!(response.status(), Status::Ok);
-        let body_string = response.body_string().expect("Expected a body");
-        //let body = serde_json::from_str(body_string.as_str());
-        //assert_eq!(, Some("Hello, world!".into()));
-    }
 }
